refactor: WinnerNumber 도메인에서 예외 로직 클래스로 분리
